# api/generated/client/typescript-axios/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/teams/{teamId}/roster": {
      "get": {
        "operationId": "Teams_getRoster",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "season",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{teamId}/recent-form": {
      "get": {
        "operationId": "Teams_getRecentForm",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "count",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamRecentFormDto"
                }
              }
            }
          }
        },
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{teamId}/bullpen": {
      "get": {
        "operationId": "Teams_getBullpen",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamBullpenDto"
                }
              }
            }
          }
        },
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{teamId}/transactions": {
      "get": {
        "operationId": "Teams_getTransactions",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamTransactionsDto"
                }
              }
            }
          }
        },
        "tags": [
          "Teams"
        ]
      }
    },
...
    "/players/search": {
      "get": {
        "operationId": "Players_searchPlayers",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "season",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerSearchResultDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
    "/players/{mlbId}": {
      "get": {
        "operationId": "Players_getPlayer",
        "parameters": [
          {
            "name": "mlbId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "season",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
    "/players/{mlbId}/team": {
      "get": {
        "operationId": "Players_getPlayerTeam",
        "parameters": [
          {
            "name": "mlbId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerTeamDto"
                }
              }
            }
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
    "/players/{mlbId}/overview/batter": {
      "get": {
        "operationId": "Players_getBatterOverview",
        "parameters": [
          {
            "name": "mlbId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "range",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatterOverviewDto"
                }
              }
            }
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
    "/players/{mlbId}/splits": {
      "get": {
        "operationId": "Players_getPlayerSplits",
        "parameters": [
          {
            "name": "mlbId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "season",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeframe",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerSplitsDto"
                }
              }
            }
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
    "/players/{mlbId}/drilldown": {
      "get": {
        "operationId": "Players_getPlayerDrilldown",
        "parameters": [
          {
            "name": "mlbId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "season",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerDrilldownDto"
                }
              }
            }
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
    "/players/{mlbId}/pitching": {
      "get": {
        "operationId": "Players_getPlayerPitching",
        "parameters": [
          {
            "name": "mlbId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "season",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerPitchingDto"
                }
              }
            }
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
    "/players/{batterId}/vs/{pitcherId}": {
      "get": {
        "operationId": "Players_getVsPlayer",
        "parameters": [
          {
            "name": "batterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pitcherId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VsPlayerDto"
                }
              }
            }
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
...
  },
  "info": {
    "title": "Baseball Realtime API",
    "description": "API for baseball-realtime",
    "version": "1.0.44",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "RecentFormGameDto": {
        "type": "object",
        "properties": {
          "providerGameId": {
            "type": "string"
          },
          "gameDate": {
            "type": "string"
          },
          "scored": {
            "type": "number"
          },
          "allowed": {
            "type": "number"
          },
          "result": {
            "type": "string",
            "enum": [
              "W",
              "L"
            ]
          }
        },
        "required": [
          "providerGameId",
          "gameDate",
          "scored",
          "allowed",
          "result"
        ]
      },
      "TeamRecentFormDto": {
        "type": "object",
        "properties": {
          "wins": {
            "type": "number"
          },
          "losses": {
            "type": "number"
          },
          "runsPerGame": {
            "type": "number"
          },
          "teamEra": {
            "type": "number",
            "nullable": true
          },
          "bullpenEra": {
            "type": "number",
            "nullable": true
          },
          "homeRuns": {
            "type": "number"
          },
          "strikeouts": {
            "type": "number"
          },
          "walks": {
            "type": "number"
          },
          "games": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentFormGameDto"
            }
          }
        },
        "required": [
          "wins",
          "losses",
          "runsPerGame",
          "homeRuns",
          "strikeouts",
          "walks",
          "games"
        ]
      },
      "BullpenPitcherDto": {
        "type": "object",
        "properties": {
          "mlbId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "hand": {
            "type": "string",
            "enum": [
              "L",
              "R"
            ]
          },
          "evidence": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "ready",
              "available",
              "rest"
            ]
          }
        },
        "required": [
          "mlbId",
          "name",
          "hand",
          "evidence",
          "state"
        ]
      },
      "TeamBullpenDto": {
        "type": "object",
        "properties": {
          "availableCount": {
            "type": "number"
          },
          "totalCount": {
            "type": "number"
          },
          "pitchers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BullpenPitcherDto"
            }
          }
        },
        "required": [
          "availableCount",
          "totalCount",
          "pitchers"
        ]
      },
      "TransactionEntryDto": {
        "type": "object",
        "properties": {
          "mlbId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "position": {
            "type": "string",
            "nullable": true
          },
          "jerseyNumber": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "injury",
              "roster",
              "trade",
              "signing"
            ]
          },
          "direction": {
            "type": "string",
            "enum": [
              "in",
              "out"
            ]
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "mlbId",
          "name",
          "date",
          "type",
          "direction",
          "description"
        ]
      },
      "TransactionsHeroDto": {
        "type": "object",
        "properties": {
          "activeRosterCount": {
            "type": "number"
          },
          "activeRosterLimit": {
            "type": "number"
          },
          "fortyManCount": {
            "type": "number"
          },
          "fortyManOpen": {
            "type": "number"
          },
          "ilCount": {
            "type": "number"
          },
          "il60Count": {
            "type": "number"
          },
          "movesThisMonth": {
            "type": "number"
          }
        },
        "required": [
          "activeRosterCount",
          "activeRosterLimit",
          "fortyManCount",
          "fortyManOpen",
          "ilCount",
          "il60Count",
          "movesThisMonth"
        ]
      },
      "TeamTransactionsDto": {
        "type": "object",
        "properties": {
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionEntryDto"
            }
          },
          "hero": {
            "$ref": "#/components/schemas/TransactionsHeroDto"
          },
          "totalMoves": {
            "type": "number"
          }
        },
        "required": [
          "transactions",
          "hero",
          "totalMoves"
        ]
      },
...
      "PlayerSearchResultDto": {
        "type": "object",
        "properties": {
          "mlbId": {
            "type": "number",
            "example": 665742
          },
          "name": {
            "type": "string",
            "example": "Yordan Alvarez"
          },
          "position": {
            "type": "string",
            "example": "DH"
          },
          "teamAbbr": {
            "type": "string",
            "example": "HOU"
          },
          "teamId": {
            "type": "number",
            "example": 117
          }
        },
        "required": [
          "mlbId",
          "name",
          "position",
          "teamAbbr",
          "teamId"
        ]
      },
      "PlayerTeamDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "false when the upstream people lookup failed"
          },
          "mlbId": {
            "type": "number",
            "example": 665161
          },
          "teamId": {
            "type": "number",
            "nullable": true,
            "example": 117
          },
          "status": {
            "type": "number",
            "description": "Upstream HTTP status, only when ok is false"
          },
          "currentTeamLink": {
            "type": "string",
            "nullable": true,
            "example": "/api/v1/teams/117"
          }
        },
        "required": [
          "ok",
          "mlbId"
        ]
      },
      "BatterOverviewHeadlineDto": {
        "type": "object",
        "properties": {
          "battingAverage": {
            "type": "string",
            "description": "The player's batting average for the season."
          },
          "onBasePercentage": {
            "type": "string",
            "description": "The player's on-base percentage for the season."
          },
          "sluggingPercentage": {
            "type": "string",
            "description": "The player's slugging percentage for the season."
          },
          "onBasePlusSlugging": {
            "type": "string",
            "description": "The player's OPS for the season."
          },
          "homeRuns": {
            "type": "number",
            "description": "The player's home runs for the season."
          },
          "runsBattedIn": {
            "type": "number",
            "description": "The player's runs batted in for the season."
          }
        },
        "required": [
          "battingAverage",
          "onBasePercentage",
          "sluggingPercentage",
          "onBasePlusSlugging",
          "homeRuns",
          "runsBattedIn"
        ]
      },
      "BatterOverviewSecondaryDto": {
        "type": "object",
        "properties": {
          "games": {
            "type": "number",
            "description": "The number of games played."
          },
          "atBats": {
            "type": "number",
            "description": "The number of at-bats."
          },
          "runs": {
            "type": "number",
            "description": "The number of runs scored."
          },
          "hits": {
            "type": "number",
            "description": "The number of hits."
          },
          "doubles": {
            "type": "number",
            "description": "The number of doubles."
          },
          "triples": {
            "type": "number",
            "description": "The number of triples."
          },
          "walks": {
            "type": "number",
            "description": "The number of walks."
          },
          "strikeouts": {
            "type": "number",
            "description": "The number of strikeouts."
          },
          "stolenBases": {
            "type": "number",
            "description": "The number of stolen bases."
          }
        },
        "required": [
          "games",
          "atBats",
          "runs",
          "hits",
          "doubles",
          "triples",
          "walks",
          "strikeouts",
          "stolenBases"
        ]
      },
...
      "BatterOverviewTodayDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "description": "The label for today's stats."
          },
          "statLine": {
            "type": "string",
            "description": "The stat line for today's stats."
          },
          "isLive": {
            "type": "boolean",
            "description": "Whether the player is currently in a live game."
          },
          "plateAppearances": {
            "type": "number",
            "nullable": true,
            "example": 4
          },
          "atBats": {
            "type": "number",
            "nullable": true,
            "example": 3
          },
          "hits": {
            "type": "number",
            "nullable": true,
            "example": 1
          },
          "homeRuns": {
            "type": "number",
            "nullable": true,
            "example": 1
          },
          "rbi": {
            "type": "number",
            "nullable": true,
            "example": 2
          },
          "walks": {
            "type": "number",
            "nullable": true,
            "example": 1
          },
          "strikeouts": {
            "type": "number",
            "nullable": true,
            "example": 1
          },
          "avg": {
            "type": "string",
            "nullable": true,
            "example": ".333"
          },
          "gameStatus": {
            "type": "string",
            "nullable": true,
            "example": "live"
          },
          "opponent": {
            "type": "string",
            "nullable": true,
            "example": "NYY"
          },
          "gameId": {
            "type": "string",
            "nullable": true,
            "example": "748531"
          },
          "playerState": {
            "type": "string",
            "nullable": true,
            "enum": [
              "atBat",
              "onDeck",
              "inTheHole",
              "idle"
            ],
            "example": "atBat"
          },
          "lastGame": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/BatterLastGameDto"
              }
            ]
          }
        },
        "required": [
          "label",
          "statLine",
          "isLive"
        ]
      },
      "BatterOverviewDto": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string",
            "description": "The player's unique identifier."
          },
          "season": {
            "type": "number",
            "description": "The season for which the stats are displayed."
          },
          "headline": {
            "description": "The player's headline stats for the season.",
            "allOf": [
              {
                "$ref": "#/components/schemas/BatterOverviewHeadlineDto"
              }
            ]
          },
          "secondary": {
            "description": "The player's secondary stats for the season.",
            "allOf": [
              {
                "$ref": "#/components/schemas/BatterOverviewSecondaryDto"
              }
            ]
          },
          "today": {
            "description": "The player's stats for today.",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/BatterOverviewTodayDto"
              }
            ]
          }
        },
        "required": [
          "playerId",
          "season",
          "headline",
          "secondary",
          "today"
        ]
      },
...
      "PlayerSplitsDto": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string",
            "example": "592450"
          },
          "season": {
            "type": "number",
            "example": 2025
          },
          "timeframe": {
            "type": "string",
            "enum": [
              "season",
              "career"
            ],
            "example": "season"
          },
          "splits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SplitRowDto"
            }
          }
        },
        "required": [
          "playerId",
          "season",
          "timeframe",
          "splits"
        ]
      },
...
      "PlayerDrilldownDto": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string",
            "example": "592450"
          },
          "season": {
            "type": "number",
            "example": 2025
          },
          "isPitcher": {
            "type": "boolean",
            "example": false
          },
          "gameLog": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameLogRowDto"
            }
          },
          "career": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CareerRowDto"
            }
          },
          "vsTeam": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VsTeamRowDto"
            }
          },
          "postseason": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CareerRowDto"
            }
          }
        },
        "required": [
          "playerId",
          "season",
          "isPitcher",
          "gameLog",
          "career",
          "vsTeam",
          "postseason"
        ]
      },
...
      "PitcherSeasonTotalsDto": {
        "type": "object",
        "properties": {
          "wins": {
            "type": "number",
            "nullable": true,
            "example": 5
          },
          "losses": {
            "type": "number",
            "nullable": true,
            "example": 2
          },
          "inningsPitched": {
            "type": "string",
            "nullable": true,
            "example": "76.1"
          },
          "era": {
            "type": "string",
            "nullable": true,
            "example": "3.18"
          },
          "whip": {
            "type": "string",
            "nullable": true,
            "example": "1.09"
          },
          "strikeOuts": {
            "type": "number",
            "nullable": true,
            "example": 72
          }
        }
      },
      "PlayerPitchingDto": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string",
            "example": "592450"
          },
          "season": {
            "type": "number",
            "example": 2025
          },
          "arsenal": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PitchArsenalRowDto"
            }
          },
          "splits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PitcherSplitRowDto"
            }
          },
          "seasonTotals": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/PitcherSeasonTotalsDto"
              }
            ]
          }
        },
        "required": [
          "playerId",
          "season",
          "arsenal",
          "splits"
        ]
      },
      "VsPlayerDto": {
        "type": "object",
        "properties": {
          "batterId": {
            "type": "number",
            "description": "Batter MLB ID",
            "example": 660670
          },
          "pitcherId": {
            "type": "number",
            "description": "Pitcher MLB ID",
            "example": 592789
          },
          "ab": {
            "type": "number",
            "description": "Career at-bats in this matchup",
            "example": 12
          },
          "h": {
            "type": "number",
            "description": "Career hits in this matchup",
            "example": 4
          },
          "hr": {
            "type": "number",
            "description": "Career home runs in this matchup",
            "example": 1
          },
          "bb": {
            "type": "number",
            "description": "Career walks in this matchup",
            "example": 0
          },
          "k": {
            "type": "number",
            "description": "Career strikeouts in this matchup",
            "example": 3
          },
          "avg": {
            "type": "string",
            "description": "Career batting average, e.g. \".333\"",
            "example": ".333",
            "nullable": true
          },
          "pa": {
            "type": "number",
            "description": "Career plate appearances in this matchup",
            "example": 15
          },
          "doubles": {
            "type": "number",
            "description": "Career doubles in this matchup",
            "example": 1
          },
          "triples": {
            "type": "number",
            "description": "Career triples in this matchup",
            "example": 0
          },
          "rbi": {
            "type": "number",
            "description": "Career RBI in this matchup",
            "example": 2
          }
        },
        "required": [
          "batterId",
          "pitcherId",
          "ab",
          "h",
          "hr",
          "bb",
          "k",
          "pa",
          "doubles",
          "triples",
          "rbi"
        ]
      },
...
    }
  }
}